Servicio básico de Asistente, falta combinar las templates

+++ src/main/java/com/ufps/conferencias/Repository/AsistenteRepository.java
@@ -9,6 +9,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AsistenteRepository extends JpaRepository <Asistente, Integer>{
+public interface AsistenteRepository extends JpaRepository <Asistente, Long>{
     
 }
